Modified the CAT operator to avoid the code(loop) repetition
